refactor(epaper_spi): extract shared 4bpp base class for inkplate 6color
inkplate 6color duplicated spectra e6's transfer_data, fill, clear and
draw_pixel_at verbatim. Move that logic into a new EPaper4bpp
intermediate base class; concrete panels now only supply their color
mapping and IC command sequences.

Also add color_to_bwyrgb (6 color) and color_to_bwyrgbo (7 color)
helpers to colorconv.h, matching the corner-mapping logic already used
by inkplate 6color and spectra e6, so the base class and color mapping
can be reused for spectra e6 in a follow-up once it's been verified on
hardware.
